fix issue #39 - add local installation and disabling of login file

+++ Makefile
@@ -1,6 +1,6 @@
 #-----------------------------------------------------------------------
 # File:         Makefile
-# Version:      2.0.1
+# Version:      2.1.2
 # Licence:      GPL 2
 # 
 # Description:  Makefile to install, uninstall Fvwm-Nightshade and create
@@ -8,7 +8,7 @@
 # 
 # Author:       Thomas Funk <t.funk@web.de>     
 # Created:      09/08/2012
-# Changed:      06/24/2013
+# Changed:      07/04/2013
 #-----------------------------------------------------------------------
 
 package 	= fvwm-nightshade
@@ -24,14 +24,19 @@ deb: DESTDIR = $(pkgdir)
 
 prefix 		?= /usr/local
 deb: prefix = /usr
+absprefix	= $(abspath $(prefix))
 
-pkgprefix	= $(DESTDIR)$(prefix)
+displaymanager ?= yes
+local		?= no
+
+pkgprefix	= $(DESTDIR)$(absprefix)
 bindir 		= $(pkgprefix)/bin
 datadir 	= $(pkgprefix)/share
 mandir 		= $(datadir)/man
 man1dir 	= $(mandir)/man1
 docdir 		= $(datadir)/doc
 localedir	= $(datadir)/locale
+fnsuserdir	= ~/.$(package)
 
 pkgdatadir 	= $(datadir)/$(package)
 pkgdocdir 	= $(docdir)/$(package)
@@ -45,12 +50,12 @@ fns_docdirs		= $(shell find doc/ -type d|sort -r|cut -d'/' -f 2-)
 fns_files 		= $(shell find $(package) -type f|cut -d'/' -f 2-)
 fns_directories = $(shell find $(package)/ -type d|sort -r|cut -d'/' -f 2-)
 fns_mofiles 	= $(shell ls -1 po |grep ".mo")
+id				= $(shell id -un)
 
 fvwm_path	?= $(DESTDIR)/usr/share/fvwm
 
 all:
 	@echo "There is nothing to compile."
-
 
 dist: $(distdir).tar.gz
 
@@ -76,15 +81,32 @@ build-install-list:
 	echo "Build install list 'fns-install_$(version).lst' for Fvwm-Nightshade $(version)"
 	echo "-> FvwmScripts"
 	if test -z "$(DESTDIR)"; then \
-		if test -d "$(fvwm_path)"; then \
+		if test "$(local)" = "yes"; then \
 			for file in $(fns_fvwmscripts); do \
-				echo $(fvwm_path)/$$file >> ./fns-install_$(version).lst; \
+				if test "$${file#*FvwmScript}" != "$$file"; then \
+					echo $(fnsuserdir)/scripts/$$file >> ./fns-install_$(version).lst; \
+				else \
+					echo $(fnsuserdir)/$$file >> ./fns-install_$(version).lst; \
+				fi; \
 			done; \
 		else \
-			echo "Fvwm isn't installed in $(fvwm_path)"; \
-			echo "Please set fvwm_path=<path_to_fvwm> and rerun make install."; \
-			rm -f ./fns-install_$(version).lst; \
-			exit 2; \
+			if test "$(id)" = "root"; then \
+				if test -d "$(fvwm_path)"; then \
+					for file in $(fns_fvwmscripts); do \
+						echo $(fvwm_path)/$$file >> ./fns-install_$(version).lst; \
+					done; \
+				else \
+					echo "Fvwm isn't installed in $(fvwm_path)"; \
+					echo "Please set fvwm_path=<path_to_fvwm> and rerun make install."; \
+					rm -f ./fns-install_$(version).lst; \
+					exit 2; \
+				fi; \
+			else \
+				echo "Can't install FvwmScripts later in $(fvwm_path) because you are not root."; \
+				echo "If you want install Fvwm-Nightshade locally use 'local=yes' in the make call."; \
+				rm -f ./fns-install_$(version).lst; \
+				exit 3; \
+			fi; \
 		fi; \
 	else \
 		for file in $(fns_fvwmscripts); do \
@@ -93,12 +115,34 @@ build-install-list:
 	fi
 
 	echo "-> Executables"
-	for file in $(fns_executables); do \
-		echo $(bindir)/$$file >> ./fns-install_$(version).lst; \
-	done
+	ok=0; \
+	slashes=`echo $(bindir) | awk '{print gsub("/","")}'|xargs expr 1 +`; \
+	while test "$$slashes" -gt "1"; do \
+		part=`echo $(bindir) |cut -d'/' -f-$$slashes`; \
+		if test -d "$$part" && test -w "$$part"; then \
+			ok=1; \
+			break; \
+		else \
+			slashes=`expr $$slashes - 1`; \
+		fi; \
+	done; \
+	if (test "$(local)" = "yes" && test "$$ok" = "1") || test "$(id)" = "root"; then \
+		for file in $(fns_executables); do \
+			echo $(bindir)/$$file >> ./fns-install_$(version).lst; \
+		done; \
+		echo $(bindir) >> ./fns-install_$(version).lst; \
+	else \
+		echo "Can't install FNS executables later in $(bindir) because you are not root."; \
+		echo "If you want install Fvwm-Nightshade locally use 'prefix=<path>' in the make call"; \
+		echo "additionally to 'local=yes' you have permissions to install it e.g. ~/local."; \
+		rm -f ./fns-install_$(version).lst; \
+		exit 3; \
+	fi; \
 
-	echo "-> Login file"
-	echo /usr/share/xsessions/fvwm-nightshade.desktop >> ./fns-install_$(version).lst
+	if test "$(displaymanager)" = "yes" && test "$(id)" = "root"; then \
+		echo "-> Login file"; \
+		echo /usr/share/xsessions/fvwm-nightshade.desktop >> ./fns-install_$(version).lst; \
+	fi
 
 	echo "-> System files"
 	for file in $(fns_files); do \
@@ -121,48 +165,73 @@ build-install-list:
 		echo $(pkgdocdir)/$$file >> ./fns-install_$(version).lst; \
 	done
 	echo $(pkgdocdir)/fns-install_$(version).lst >> ./fns-install_$(version).lst
+	if test "$(displaymanager)" = "no" || (test "$(local)" = "yes" && test "$(id)" != "root"); then \
+		echo $(pkgdocdir)/fvwm-nightshade.desktop-example >> ./fns-install_$(version).lst; \
+	fi
+	echo "-> Template files"
+	for file in $(fns_templates); do \
+		echo $(pkgdocdir)/templates/$$file >> ./fns-install_$(version).lst; \
+	done
 	
 	echo "-> Documentation directories"
 	for directory in $(fns_docdirs); do \
 		echo $(pkgdocdir)/$$directory >> ./fns-install_$(version).lst; \
 	done
-	
-	echo "-> Template files"
-	for file in $(fns_templates); do \
-		echo $(pkgdocdir)/templates/$$file >> ./fns-install_$(version).lst; \
-	done
 	echo $(pkgdocdir)/templates >> ./fns-install_$(version).lst;
 	echo $(pkgdocdir) >> ./fns-install_$(version).lst
+	echo $(docdir) >> ./fns-install_$(version).lst
 
 	echo "-> Manpage files"
 	for file in $(fns_manpages); do \
 		echo $(man1dir)/$$file >> ./fns-install_$(version).lst; \
 	done
+	echo $(man1dir) >> ./fns-install_$(version).lst
+	echo $(mandir) >> ./fns-install_$(version).lst
+	echo $(datadir) >> ./fns-install_$(version).lst
+	echo $(pkgprefix) >> ./fns-install_$(version).lst
 
 	echo "-> Localization files"
 	for file in $(fns_mofiles); do \
 		basename=$${file%%.*}; \
 		extensions=$${file#*.}; \
 		lang=$${extensions%%.*}; \
-		echo $(localedir)/$$lang/LC_MESSAGES/$$basename.mo >> ./fns-install_$(version).lst; \
+		if test "$(local)" = "yes" && test "$(id)" != "root"; then \
+			echo $(fnsuserdir)/locale/$$lang/LC_MESSAGES/$$basename.mo >> ./fns-install_$(version).lst; \
+		else \
+			echo $(localedir)/$$lang/LC_MESSAGES/$$basename.mo >> ./fns-install_$(version).lst; \
+		fi; \
 	done
 
 	echo "Done"
 	echo
-	
-install: build-install-list
+
+install: build-install-list dist-install
+
+dist-install:
 	echo "Installing Fvwm-Nightshade $(version) to $(pkgprefix)"
 	echo "-> Install FvwmScripts"
 	if test -z "$(DESTDIR)"; then \
-		if test -d "$(fvwm_path)"; then \
+		if test "$(local)" = "yes"; then \
+			install -d $(fnsuserdir)/scripts; \
 			for file in $(fns_fvwmscripts); do \
-				install -m 644 fvwm/$$file  $(fvwm_path); \
+				if test "$${file#*FvwmScript}" != "$$file"; then \
+					install -m 644 fvwm/$$file $(fnsuserdir)/scripts/; \
+				else \
+					install -m 644 fvwm/$$file $(fnsuserdir); \
+				fi; \
 			done; \
 		else \
-			echo "Fvwm isn't installed in $(fvwm_path)"; \
-			echo "Please set fvwm_path=<path_to_fvwm> and rerun make install."; \
-			rm -f ./fns-install_$(version).lst; \
-			exit 2; \
+			if test "$(id)" = "root"; then \
+				if test -d "$(fvwm_path)"; then \
+					for file in $(fns_fvwmscripts); do \
+						install -m 644 fvwm/$$file  $(fvwm_path); \
+					done; \
+				else \
+					exit 2; \
+				fi; \
+			else \
+				exit 3; \
+			fi; \
 		fi; \
 	else \
 		install -d $(fvwm_path); \
@@ -177,25 +246,33 @@ install: build-install-list
 		install -m 755 bin/$$file $(bindir); \
 	done
 
-	echo "-> Install login file"
-	install -d $(DESTDIR)/usr/share/xsessions
-	install -m 644 system/fvwm-nightshade.desktop $(DESTDIR)/usr/share/xsessions/
+	if test -n "$(DESTDIR)" || (test "$(displaymanager)" = "yes" && test "$(id)" = "root"); then \
+		echo "-> Install login file"; \
+		install -d $(DESTDIR)/usr/share/xsessions; \
+		install -m 644 system/fvwm-nightshade.desktop $(DESTDIR)/usr/share/xsessions/;\
+	fi
 
 	echo "-> Install system files"
 	install -d $(datadir)
 	cp -r $(package) $(pkgdatadir)
 	
-	echo "-> Install documentation, Readmes and templates"
+	echo "-> Install documentation, Readmes, examples and templates"
 	install -d $(pkgdocdir)
 	cp -r doc/* $(pkgdocdir)
 
 	for file in AUTHORS ChangeLog COPYING NEWS README INSTALL; do \
 		install -m 644 $$file $(pkgdocdir); \
 	done
+
+	if test "$(displaymanager)" = "no" || (test "$(local)" = "yes" && test "$(id)" != "root"); then \
+		install -m 644 system/fvwm-nightshade.desktop $(pkgdocdir)/fvwm-nightshade.desktop-example; \
+	fi
 	
 	cp -r templates $(pkgdocdir)
 
-	install -m 644 fns-install_$(version).lst $(pkgdocdir)
+	if test -z "$(DESTDIR)"; then \
+		install -m 644 fns-install_$(version).lst $(pkgdocdir); \
+	fi
 	
 	echo "-> Install manpage files"
 	install -d $(man1dir)
@@ -208,8 +285,13 @@ install: build-install-list
 		basename=$${file%%.*}; \
 		extensions=$${file#*.}; \
 		lang=$${extensions%%.*}; \
-		install -d $(localedir)/$$lang/LC_MESSAGES; \
-		install -m 644 po/$$file $(localedir)/$$lang/LC_MESSAGES/$$basename.mo; \
+		if test "$(local)" = "yes" && test "$(id)" != "root"; then \
+			install -d $(fnsuserdir)/locale/$$lang/LC_MESSAGES; \
+			install -m 644 po/$$file $(fnsuserdir)/locale/$$lang/LC_MESSAGES/$$basename.mo; \
+		else \
+			install -d $(localedir)/$$lang/LC_MESSAGES; \
+			install -m 644 po/$$file $(localedir)/$$lang/LC_MESSAGES/$$basename.mo; \
+		fi; \
 	done
 	
 	if test -z "$(DESTDIR)"; then \
@@ -230,11 +312,11 @@ uninstall:
 				echo "copy $(pkgdocdir)/fns-install_$$inst_fns_version.lst to `pwd`"; \
 				for path in `cat fns-install_$$inst_fns_version.lst` ; do \
 					if test -f "$$path"; then \
-						rm -f $$path; \
 						echo "remove file: $$path"; \
+						rm -f $$path; \
 					elif test -d "$$path"; then \
-						rmdir $$path; \
 						echo "remove directory: $$path"; \
+						rmdir $$path; \
 					else \
 						echo "not found: $$path"; \
 					fi; \
@@ -245,55 +327,14 @@ uninstall:
 			else \
 				echo "Install log ($$inst_fns_version) and config version ($$fns_version) differs!"; \
 				echo "Try 'make uninstall-alternative' to remove files/directories otherwise."; \
-				echo "But then file/directory fragments can be remained."; \
+				echo "But then file/directory fragments can remain."; \
 				exit 2; \
 			fi; \
 		else \
 			echo "Can't find fns-install log under $(pkgdocdir)! Wrong installation path."; \
 			echo "Please set correct prefix=<other_dir> and rerun make uninstall."; \
 			echo "Or use 'make uninstall-alternative' to remove files/directories"; \
-			echo "otherwise. But then file/directory fragments can be remained."; \
-			exit 2; \
-		fi; \
-	else \
-		echo "Can't find config in $(pkgdatadir)! Wrong installation path."; \
-		echo "Please set correct prefix=<other_dir> and rerun make uninstall."; \
-		exit 2; \
-	fi
-
-dry-uninstall:
-	echo "Uninstall previous version of Fvwm-Nightshade - dry run. Nothing happens."
-	if test -f "$(pkgdatadir)/config"; then\
-		fns_version=`grep ns_version $(pkgdatadir)/config |sed q |cut -d' ' -f3`; \
-		uninst_file=`find $(pkgdocdir)/fns-install_*.lst`; \
-		if test -f "$$uninst_file" && test "$$uninst_file" != ''; then \
-			inst_fns_version=`find $(pkgdocdir)/fns-install_*.lst |cut -d '_' -f2|sed 's/.lst//'`; \
-			if test "$$fns_version" = "$$inst_fns_version"; then \
-				echo "Install log version and config version identically: $$fns_version"; \
-				echo "Starting with deinstallation ..."; \
-				echo "copy $(pkgdocdir)/fns-install_$$inst_fns_version.lst to `pwd`"; \
-				for path in `cat fns-install_$$inst_fns_version.lst` ; do \
-					if test -f "$$path"; then \
-						echo "remove file: $$path"; \
-					elif test -d "$$path"; then \
-						echo "remove directory: $$path"; \
-					else \
-						echo "not found: $$path"; \
-					fi; \
-				done; \
-				echo "Dry run done.."; \
-				exit 0; \
-			else \
-				echo "Install log ($$inst_fns_version) and config version ($$fns_version) differs!"; \
-				echo "Try 'make uninstall-alternative' to remove files/directories otherwise."; \
-				echo "But then file/directory fragments can be remained."; \
-				exit 2; \
-			fi; \
-		else \
-			echo "Can't find fns-install log under $(pkgdocdir)! Wrong installation path."; \
-			echo "Please set correct prefix=<other_dir> and rerun make uninstall."; \
-			echo "Or use 'make uninstall-alternative' to remove files/directories"; \
-			echo "otherwise. But then file/directory fragments can be remained."; \
+			echo "otherwise. But then file/directory fragments can remain."; \
 			exit 2; \
 		fi; \
 	else \
@@ -308,8 +349,8 @@ uninstall-alternative:
 	if test -d "$(fvwm_path)"; then \
 		for file in $(fns_fvwmscripts) ; do \
 			if test -f "$(fvwm_path)/$$file"; then \
-				rm -f $(fvwm_path)/$$file; \
 				echo "remove $(fvwm_path)/$$file"; \
+				rm -f $(fvwm_path)/$$file; \
 			fi; \
 		done; \
 	else \
@@ -321,32 +362,34 @@ uninstall-alternative:
 	echo "-> Uninstall executables"
 	for file in $(fns_executables) ; do \
 		if test -f "$(bindir)/$$file"; then \
-			rm -f $(bindir)/$$file; \
 			echo "remove $(bindir)/$$file"; \
+			rm -f $(bindir)/$$file; \
 		fi; \
 	done
 
-	echo "-> Uninstall login file"
-	-rm /usr/share/xsessions/fvwm-nightshade.desktop
-	echo "remove /usr/share/xsessions/fvwm-nightshade.desktop"
+	if test -f "/usr/share/xsessions/fvwm-nightshade.desktop"; then \
+		echo "-> Uninstall login file"; \
+		echo "remove /usr/share/xsessions/fvwm-nightshade.desktop"; \
+		-rm /usr/share/xsessions/fvwm-nightshade.desktop; \
+	fi
 
 	echo "-> Uninstall system files"
 	if test -d "$(pkgdatadir)"; then \
-		-rm -r $(pkgdatadir); \
 		echo "remove $(pkgdatadir) completelly"; \
+		-rm -r $(pkgdatadir); \
 	fi
 	
 	echo "-> Uninstall documentation"
 	if test -d "$(pkgdocdir)"; then \
-		-rm -r $(pkgdocdir); \
 		echo "remove $(pkgdocdir) completelly"; \
+		-rm -r $(pkgdocdir); \
 	fi
 
 	echo "-> Uninstall manpages"
 	for file in $(fns_manpages) ; do \
 		if test -f "$(man1dir)/$$file"; then \
-			rm -f $(man1dir)/$$file; \
 			echo "remove $(man1dir)/$$file"; \
+			rm -f $(man1dir)/$$file; \
 		fi; \
 	done
 
@@ -356,67 +399,13 @@ uninstall-alternative:
 		extensions=$${file#*.}; \
 		lang=$${extensions%%.*}; \
 		if test -f "$(localedir)/$$lang/LC_MESSAGES/$$basename.mo"; then \
-			rm -f $(localedir)/$$lang/LC_MESSAGES/$$basename.mo; \
 			echo "remove $(localedir)/$$lang/LC_MESSAGES/$$basename.mo"; \
+			rm -f $(localedir)/$$lang/LC_MESSAGES/$$basename.mo; \
 		fi; \
 	done
 
 	echo "Fvwm-Nightshade is now removed (hopefully). Only ~/.fvwm-nightshade exists."
 	echo "If you don't need it anymore remove it by hand."
-
-dry-uninstall-alternative:
-	echo "Try to uninstall previous version of Fvwm-Nightshade - dry run. Nothing happens."
-	echo "-> Uninstall FvwmScripts"
-	if test -d "$(fvwm_path)"; then \
-		for file in $(fns_fvwmscripts) ; do \
-			if test -f "$(fvwm_path)/$$file"; then \
-				echo "remove $(fvwm_path)/$$file"; \
-			fi; \
-		done; \
-	else \
-		echo "Fvwm isn't installed in $(fvwm_path)"; \
-		echo "Please set fvwm_path=<path_to_fvwm> and rerun make uninstall."; \
-		exit 2; \
-	fi
-	
-	echo "-> Uninstall executables"
-	for file in $(fns_executables) ; do \
-		if test -f "$(bindir)/$$file"; then \
-			echo "remove $(bindir)/$$file"; \
-		fi; \
-	done
-
-	echo "-> Uninstall login file"
-	echo "remove /usr/share/xsessions/fvwm-nightshade.desktop"
-
-	echo "-> Uninstall fvwm-nightshade system files"
-	if test -d "$(pkgdatadir)"; then \
-		echo "remove $(pkgdatadir) completelly"; \
-	fi
-	
-	echo "-> Uninstall documentation"
-	if test -d "$(pkgdocdir)"; then \
-		echo "remove $(pkgdocdir) completelly"; \
-	fi
-
-	echo "-> Uninstall manpages"
-	for file in $(fns_manpages) ; do \
-		if test -f "$(man1dir)/$$file"; then \
-			echo "remove $(man1dir)/$$file"; \
-		fi; \
-	done
-
-	echo "-> Uninstall localization files"
-	for file in $(fns_mofiles); do \
-		basename=$${file%%.*}; \
-		extensions=$${file#*.}; \
-		lang=$${extensions%%.*}; \
-		if test -f "$(localedir)/$$lang/LC_MESSAGES/$$basename.mo"; then \
-			echo "remove $(localedir)/$$lang/LC_MESSAGES/$$basename.mo"; \
-		fi; \
-	done
-
-	echo "Dry run done."
 
 build-deb: 
 	echo "Build debian package"
@@ -427,7 +416,7 @@ build-deb:
 	mkdir -p $(pkgdocdir)
 	cp debian/copyright $(pkgdocdir)
 
-deb: build-deb install
+deb: build-deb dist-install
 	sed -i "s/Version:/Version: $(version)/" $(pkgdir)/DEBIAN/control
 	sed -i "s/Installed-Size:/Installed-Size: `du -s fvwm-nightshade |cut -f1`/" $(pkgdir)/DEBIAN/control
 	dpkg -b $(pkgdir) ../$(package)_$(version)_all.deb
@@ -459,6 +448,6 @@ arch: prepare-arch
 	mv $(pkgdir)/*.xz ../
 	rm -rf $(pkgdir)
 	
-.PHONY: dist distcheck install uninstall dry-uninstall uninstall-alternative dry-uninstall-alternative deb rpm arch
-.SILENT: FORCE dist install uninstall build-deb deb rpm prepare-rpm arch prepare-arch build-install-list dry-uninstall uninstall-alternative dry-uninstall-alternative
+.PHONY: dist distcheck install uninstall uninstall-alternative deb rpm arch
+.SILENT: FORCE dist install uninstall build-deb deb rpm prepare-rpm arch prepare-arch build-install-list uninstall-alternative dist-install
